backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
supabase==2.10.0
anthropic==0.40.0
python-multipart==0.0.20
python-dotenv==1.0.1
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
